release and add mp3s

+++ datagather/src/main/kotlin/ro/adriantosca/cipcirip/datagather/csv/OrganismMedia.kt
@@ -1,6 +1,6 @@
 package ro.adriantosca.cipcirip.model
 
 data class OrganismMedia(
-    var organismId: Long,
-    var mediaId: Long
+    var organismId: Int,
+    var mediaId: Int
 )
